+++ baybe/utils/__init__.py
@@ -10,7 +10,6 @@ import random
 import ssl
 import urllib.request
 from abc import ABC
-from copy import deepcopy
 from dataclasses import dataclass
 from functools import lru_cache
 from pathlib import Path
@@ -35,7 +34,6 @@ import torch
 from attrs import cmp_using
 from joblib import Memory
 from mordred import Calculator, descriptors
-from pydantic import BaseModel as PydanticBaseModel
 from rdkit import Chem, RDLogger
 from rdkit.Chem.AllChem import GetMorganFingerprintAsBitVect
 from torch import Tensor
@@ -617,16 +615,6 @@ def df_uncorrelated_features(
     return data
 
 
-class StrictValidationError(Exception):
-    """
-    This class is used as a workaround that can be used to stop pydantic from continuing
-    validating other members when an earlier validation failed.
-
-    The issue is described here:
-    https://github.com/pydantic/pydantic/issues/3915
-    """
-
-
 class NotEnoughPointsLeftError(Exception):
     """An exception raised when more recommendations are requested than there are
     viable parameter configurations left in the search space."""
@@ -728,46 +716,6 @@ def set_random_seed(seed: int) -> None:
     torch.manual_seed(seed)
     random.seed(seed)
     np.random.seed(seed)
-
-
-class BaseModel(PydanticBaseModel):
-    """Pydantic model with some basic settings used for various BayBE components."""
-
-    class Config:  # pylint: disable=missing-class-docstring
-        arbitrary_types_allowed = True
-        json_encoders = {
-            pd.DataFrame: lambda x: x.to_dict(orient="list"),
-        }
-
-
-class ABCBaseModel(BaseModel):
-    """Pydantic model for class hierarchies with "type" logic."""
-
-    # TODO: This is only a temporary workaround. The deserialization of subclasses
-    #   will be refactored once all class structures have been cleaned up.
-
-    @classmethod
-    def parse_obj(cls, obj):
-        return cls._convert_to_real_type_(obj)
-
-    @classmethod
-    def _convert_to_real_type_(cls, data):
-
-        try:
-            data = deepcopy(data)
-            data_type = data.pop("type")
-        except AttributeError:
-            return data
-        except KeyError:
-            return cls(**data)
-
-        sub = cls.SUBCLASSES.get(data_type)
-
-        return sub(**data)
-
-    @classmethod
-    def __get_validators__(cls):
-        yield cls._convert_to_real_type_
 
 
 def unstructure_base(base):